requirements.txt
Django==1.11.17
djangorestframework==3.8.2
sqlalchemy==1.3.17
alembic==0.9.9
psycopg2-binary